Fix UWP CI build
NuGet.config file should be in the repository for the AppCenter Build.
When the build configuration is saved, the build configuration builder
scans the repo for NuGet.config file and if it exists, it will setup
the build pipeline accordingly.
Note, that this decision is happening at the build configuration "save"
stage. The build runner doesn't check if NuGet config is present in the
project. So post-clone script would not do the job, and thus we still
need the file in our repo.
Having an empty NuGet.file in repo and replacing its content by
the post-clone script will do the job.