Fix checking pending logs on SDK start
This issue was found when running on Windows Server machine.
Channel was stuck in dead lock while trying to check Pending logs
on start and at the same time trying to perform other async operations.

Removing checkPendingLogs from Channel's ctor and moving it
to ChannelGroup on adding a channel solves the problem. This solution
is also consistent with Android SDK implementation.