Fix a deadlock that occurs during the services initialization
Currently, when the task responsible for releasing the lock acquired during the `Channel` instantiation is not quickly executed and remains in the scheduler queue while there are other tasks executing concurrently, it can lead to deadlocks.
Change: make the task that initializes  `_pendingCountLog` available to private methods  so that it can `awaited` on when needing to evaluate and / or update the field.