Don't use RegexOptions.Compiled at all
Normally AppCenter.Start should be called once.
In case the AppCenter.Start is called multiple times,
it won't improve the performance significantly. Moreover,
it is ignored on Xamarin.Android and mono, and gets much longer
to execute on net6+ for Android. See benchmark results here:
https://github.com/jonathanpeppers/BenchmarkDotNet-Android

There was a discussion on whether this flag should be set here:
https://github.com/microsoft/appcenter-sdk-dotnet/pull/1253#pullrequestreview-330047060

Based on the discussion, the final point was that Xamarin Forms may
call onStart multiple times. The documentation says:

> On Android, the OnStart method will be called on rotation as well as when the application first starts, if the main activity lacks ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation in the [Activity()] attribute.

However, since the flag is ignored on Xamarin.Android, it doesn't help us with this scenario anyway.